Avoid warnings in saturating casts test

+++ test/correctness/saturating_casts.cpp
@@ -82,11 +82,11 @@ void test_saturating() {
                                                        (int64_t)target_min),
                                               (int64_t)target_max);
 
-            if (simpler_correct_result != correct_result) {
+            if (simpler_correct_result != (int64_t)correct_result) {
                 std::cout << "Simpler verification failed for index " << i << " correct_result is " << correct_result << " correct_result casted to int64_t is " << (int64_t)correct_result << " simpler_correct_result is " << simpler_correct_result << "\n";
                 std::cout << "in(i) " << in(i) << " target_min " << target_min << " target_max " << target_max << "\n";
             }
-            assert(simpler_correct_result == correct_result);
+            assert(simpler_correct_result == (int64_t)correct_result);
         }
 
         if (result(i) != correct_result) {
@@ -180,11 +180,16 @@ void test_concise(cast_maker_t cast_maker, bool saturating) {
                                                     (int64_t)target_max);
                 }
 
-                if (simpler_correct_result != correct_result) {
-                    std::cout << "Simpler verification failed for index " << i << " correct_result is " << correct_result << " correct_result casted to int64_t is " << (int64_t)correct_result << " simpler_correct_result is " << simpler_correct_result << "\n";
-                    std::cout << "in(i) " << in(i) << " target_min " << target_min << " target_max " << target_max << "\n";
+                if (simpler_correct_result != (int64_t)correct_result) {
+                    std::cout << "Simpler verification failed for index " << i
+                              << " correct_result is " << correct_result
+                              << " correct_result casted to int64_t is " << (int64_t)correct_result
+                              << " simpler_correct_result is " << simpler_correct_result << "\n";
+                    std::cout << "in(i) " << in(i)
+                              << " target_min " << target_min
+                              << " target_max " << target_max << "\n";
                 }
-                assert(simpler_correct_result == correct_result);
+                assert(simpler_correct_result == (int64_t)correct_result);
             }
 
         } else {
@@ -192,7 +197,11 @@ void test_concise(cast_maker_t cast_maker, bool saturating) {
         }
 
         if (result(i) != correct_result) {
-            std::cout << "Match failure at index " << i << " got " << result(i) << " expected " << correct_result << " for input " << in(i) << (saturating ? " saturating" : " nonsaturating") << std::endl;
+            std::cout << "Match failure at index " << i
+                      << " got " << result(i)
+                      << " expected " << correct_result
+                      << " for input " << in(i)
+                      << (saturating ? " saturating" : " nonsaturating") << std::endl;
         }
 
         assert(result(i) == correct_result);
